🐛 fix: Erro ao excluir despesa
Correção ao excluir despesa referente a pagamento de fatura.
Estava utilizando o ID da fatura ao invés do ID da transação.

+++ lib/features/transactions/domain/use_cases/expense_delete.dart
@@ -5,6 +5,7 @@ import 'package:finan_master_app/features/transactions/domain/entities/expense_e
 import 'package:finan_master_app/features/transactions/domain/repositories/i_expense_repository.dart';
 import 'package:finan_master_app/features/transactions/domain/use_cases/i_expense_delete.dart';
 import 'package:finan_master_app/shared/domain/repositories/i_local_db_transaction_repository.dart';
+import 'package:finan_master_app/shared/infra/data_sources/database_local/i_database_local_transaction.dart';
 
 class ExpenseDelete implements IExpenseDelete {
   final IExpenseRepository _repository;
@@ -23,19 +24,21 @@ class ExpenseDelete implements IExpenseDelete {
         _localDBTransactionRepository = localDBTransactionRepository;
 
   @override
-  Future<void> delete(ExpenseEntity entity) async {
+  Future<void> delete(ExpenseEntity entity, {ITransactionExecutor? txn}) async {
+    if (txn == null) {
+      return await _localDBTransactionRepository.openTransaction((newTxn) => delete(entity, txn: newTxn));
+    }
+
     CreditCardTransactionEntity? creditCardTransaction;
 
     if (entity.idCreditCardTransaction != null) {
-      creditCardTransaction = await _creditCardTransactionRepository.findById(entity.idCreditCardTransaction!);
+      creditCardTransaction = await _creditCardTransactionRepository.findById(entity.idCreditCardTransaction!, txn: txn);
     }
 
-    await _localDBTransactionRepository.openTransaction((txn) async {
-      await Future.wait([
-        _repository.delete(entity, txn: txn),
-        if (creditCardTransaction != null) _creditCardTransactionRepository.delete(creditCardTransaction, txn: txn),
-        _statementRepository.deleteByIdExpense(entity.id, txn: txn),
-      ]);
-    });
+    await Future.wait([
+      _repository.delete(entity, txn: txn),
+      if (creditCardTransaction != null) _creditCardTransactionRepository.delete(creditCardTransaction, txn: txn),
+      _statementRepository.deleteByIdExpense(entity.id, txn: txn),
+    ]);
   }
 }
