+++ lib/features/transactions/domain/use_cases/transfer_delete.dart
@@ -3,6 +3,7 @@ import 'package:finan_master_app/features/transactions/domain/entities/transfer_
 import 'package:finan_master_app/features/transactions/domain/repositories/i_transfer_repository.dart';
 import 'package:finan_master_app/features/transactions/domain/use_cases/i_transfer_delete.dart';
 import 'package:finan_master_app/shared/domain/repositories/i_local_db_transaction_repository.dart';
+import 'package:finan_master_app/shared/infra/data_sources/database_local/i_database_local_transaction.dart';
 
 class TransferDelete implements ITransferDelete {
   final ITransferRepository _repository;
@@ -18,12 +19,14 @@ class TransferDelete implements ITransferDelete {
         _localDBTransactionRepository = localDBTransactionRepository;
 
   @override
-  Future<void> delete(TransferEntity entity) async {
-    await _localDBTransactionRepository.openTransaction((txn) async {
-      await Future.wait([
-        _repository.delete(entity, txn: txn),
-        _statementRepository.deleteByIdTransfer(entity.id, txn: txn),
-      ]);
-    });
+  Future<void> delete(TransferEntity entity, {ITransactionExecutor? txn}) async {
+    if (txn == null) {
+      return await _localDBTransactionRepository.openTransaction((newTxn) => delete(entity, txn: newTxn));
+    }
+
+    await Future.wait([
+      _repository.delete(entity, txn: txn),
+      _statementRepository.deleteByIdTransfer(entity.id, txn: txn),
+    ]);
   }
 }
